refactor: move i18n code to src/core/i18n.ts and rename class to I18nResolver
- Create src/core/i18n.ts with I18nResolver class
- Move code from src/messages/tr-messages.ts to new location
- Maintain backward compatibility by re-exporting TrMessages alias
- Update all internal imports to use new I18nResolver
- Export I18nResolver from src/core/index.ts
- All tests pass (246/246)

Fixes #15
