+++ src/core/index.ts
@@ -4,5 +4,6 @@ export * from './form';
 export * from './trivule';
 export * from './validate';
 export * from './bag';
+export * from './i18n';
 
 export * from './utils/rule-executed';
