--- src/messages/index.ts
@@ -1,2 +1,0 @@
-export * from './tr-messages';
-export * from '../core/utils/input-rule';
