+++ tests/validation.test.ts
@@ -1,5 +1,5 @@
 import { TrValidation } from '../src/core/validate';
-import { InputRule } from '../src/messages';
+import { InputRule } from '../src/core/utils/input-rule';
 
 describe('TrValidation', () => {
   let trvalidation: TrValidation;
